+++ analysis/src/SharedTypes.ml
@@ -775,7 +775,7 @@ end
 
 module Completion = struct
   type kind =
-    | Module of Module.t
+    | Module of {docstring: string list; module_: Module.t}
     | Value of Types.type_expr
     | ObjLabel of Types.type_expr
     | Label of string
@@ -800,10 +800,11 @@ module Completion = struct
     kind: kind;
     detail: string option;
     typeArgContext: typeArgContext option;
+    data: (string * string) list option;
   }
 
-  let create ?typeArgContext ?(includesSnippets = false) ?insertText ~kind ~env
-      ?sortText ?deprecated ?filterText ?detail ?(docstring = []) name =
+  let create ?data ?typeArgContext ?(includesSnippets = false) ?insertText ~kind
+      ~env ?sortText ?deprecated ?filterText ?detail ?(docstring = []) name =
     {
       name;
       env;
@@ -817,6 +818,7 @@ module Completion = struct
       filterText;
       detail;
       typeArgContext;
+      data;
     }
 
   (* https://microsoft.github.io/language-server-protocol/specifications/specification-current/#textDocument_completion *)